+++ pkg/bundler/bundler.go
@@ -2,6 +2,7 @@ package bundler
 
 import (
 	"context"
+	"fmt"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -22,7 +23,7 @@ type Bundler struct {
 	name                  string
 	entryPoints           []string
 	absWorkingDir         string
-	watchPaths            []string
+	watchPaths            []*watcher.WatchPath
 	ignorePaths           []string
 	log                   abstractlogger.Logger
 	skipWatchOnEntryPoint bool
@@ -32,7 +33,7 @@ type Bundler struct {
 	fileLoaders           []string
 	mu                    sync.Mutex
 	buildResult           *api.BuildResult
-	onAfterBundle         func()
+	onAfterBundle         func() error
 }
 
 type Config struct {
@@ -41,11 +42,11 @@ type Config struct {
 	AbsWorkingDir         string
 	SkipWatchOnEntryPoint bool
 	EntryPoints           []string
-	WatchPaths            []string
+	WatchPaths            []*watcher.WatchPath
 	IgnorePaths           []string
 	OutFile               string
 	OutDir                string
-	OnAfterBundle         func()
+	OnAfterBundle         func() error
 }
 
 func NewBundler(config Config) *Bundler {
@@ -64,7 +65,7 @@ func NewBundler(config Config) *Bundler {
 	}
 }
 
-func (b *Bundler) Bundle() {
+func (b *Bundler) Bundle() error {
 	if b.buildResult != nil {
 		buildResult := b.buildResult.Rebuild()
 		b.buildResult = &buildResult
@@ -73,7 +74,7 @@ func (b *Bundler) Bundle() {
 				abstractlogger.String("bundlerName", b.name),
 				abstractlogger.Any("errors", b.buildResult.Errors),
 			)
-			return
+			return fmt.Errorf("build failed: %s, %s", b.buildResult.Errors[0].Location.LineText, b.buildResult.Errors[0].Text)
 		}
 		b.log.Debug("Build successful", abstractlogger.String("bundlerName", b.name))
 	} else {
@@ -84,13 +85,15 @@ func (b *Bundler) Bundle() {
 				abstractlogger.String("bundlerName", b.name),
 				abstractlogger.Any("errors", b.buildResult.Errors),
 			)
-			return
+			return fmt.Errorf("build failed: %s, %s", b.buildResult.Errors[0].Location.LineText, b.buildResult.Errors[0].Text)
 		}
 		b.log.Debug("Initial Build successful", abstractlogger.String("bundlerName", b.name))
 	}
 	if b.onAfterBundle != nil {
-		b.onAfterBundle()
+		return b.onAfterBundle()
 	}
+
+	return nil
 }
 
 func (b *Bundler) Watch(ctx context.Context) {
@@ -105,7 +108,7 @@ func (b *Bundler) Watch(ctx context.Context) {
 			abstractlogger.String("bundlerName", b.name),
 			abstractlogger.String("outFile", b.outFile),
 			abstractlogger.Strings("externalImports", b.externalImports),
-			abstractlogger.Strings("watchPaths", b.watchPaths),
+			abstractlogger.Any("watchPaths", b.watchPaths),
 			abstractlogger.Strings("fileLoaders", b.fileLoaders),
 		)
 		b.watch(ctx, b.buildResult.Rebuild)
@@ -175,7 +178,7 @@ func (b *Bundler) initialBuild() api.BuildResult {
 			// check if the path already exist
 			exists := false
 			for _, watchPath := range b.watchPaths {
-				if watchPath == file {
+				if watchPath.Path == file {
 					exists = true
 					break
 				}
@@ -185,7 +188,7 @@ func (b *Bundler) initialBuild() api.BuildResult {
 					// each plugin runs on a separate go routine
 					b.mu.Lock()
 					defer b.mu.Unlock()
-					b.watchPaths = append(b.watchPaths, file)
+					b.watchPaths = append(b.watchPaths, &watcher.WatchPath{Path: file})
 				} else {
 					b.log.Error("Bundler watching limit exceeded", abstractlogger.String("bundlerName", b.name), abstractlogger.Int("limit", watchFileLimit), abstractlogger.Error(err))
 				}